refactor: remove python dependency
